Documentation and prevent Infinity to be deleted

+++ Projects/Projects.rb
@@ -103,6 +103,8 @@ class Projects
     def self.destroy(project)
         uuid = project["uuid"]
         return if uuid == "20200502-141331-226084" # Guardian General Work
+        return if uuid == "44caf74675ceb79ba5cc13bafa102509369c2b53" # Inbox
+        return if uuid == "0219fd54bd5841008b18c414a5b2dea331bad1c5" # Infinity
         filepath = "#{Projects::pathToProjects()}/#{uuid}.json"
         return if !File.exists?(filepath)
         FileUtils.rm(filepath)
